backend/flight_mcp_agent/requirements.txt
modelcontextprotocol
amadeus
pydantic>=2.0.0
python-dotenv
httpx
fastmcp